Reworked MovieService.java

+++ src/main/java/com/example/kino_xp/repository/SeatRowRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeatRowRepository extends JpaRepository<SeatRow, Integer> {
+public interface SeatRowRepository extends JpaRepository<SeatRow, Long> {
 }
